requirements.txt
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.0.76
grpcio
grpcio-tools
Pillow
numpy<2.0